process file to avoid circular imports, and get pytests working
