+++ vivarium/actor/shepherd.py
@@ -2,6 +2,7 @@ from __future__ import absolute_import, division, print_function
 
 import multiprocessing as mp
 import vivarium.actor.event as event
+import vivarium.core.process
 import vivarium.core.tree
 from vivarium.actor.actor import Actor
 
@@ -81,7 +82,7 @@ class ActorShepherd(Actor):
 
         initializer = self.agent_initializers.get(agent_type, None)
         if initializer:
-            process = vivarium.core.tree.Process(
+            process = vivarium.core.process.Process(
                 target=initializer,
                 name=agent_id,
                 args=(agent_id, agent_type, actor_config))
