+++ vivarium/core/composition.py
@@ -8,13 +8,11 @@ import numpy as np
 import networkx as nx
 
 from vivarium.core.tree import (
-    Process,
-    Deriver,
     Experiment,
     update_in,
     generate_derivers,
 )
-from vivarium.core.process import deriver_library
+from vivarium.core.process import Process, Deriver
 from vivarium.core.tree import Compartment as TreeCompartment
 from vivarium.core import emitter as emit
 from vivarium.utils.dict_utils import (
