+++ vivarium/processes/Kremling2007_transport.py
@@ -7,7 +7,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 from scipy.integrate import odeint
 
-from vivarium.core.tree import Process
+from vivarium.core.process import Process
 from vivarium.utils.flux_conversion import millimolar_to_counts, counts_to_millimolar
 from vivarium.utils.make_media import Media
 from vivarium.core.composition import PROCESS_OUT_DIR
