+++ vivarium/processes/complexation.py
@@ -6,7 +6,7 @@ from arrow import StochasticSystem
 
 from vivarium.data.molecular_weight import molecular_weight
 from vivarium.utils.dict_utils import keys_list
-from vivarium.core.tree import Process
+from vivarium.core.process import Process
 from vivarium.data.chromosomes.flagella_chromosome import FlagellaChromosome
 
 chromosome = FlagellaChromosome()
