+++ vivarium/processes/degradation.py
@@ -4,7 +4,7 @@ import os
 import copy
 
 from vivarium.utils.dict_utils import keys_list
-from vivarium.core.tree import Process
+from vivarium.core.process import Process
 from vivarium.core.composition import (
     simulate_process,
     plot_simulation_output,
