+++ vivarium/processes/meta_division.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, division, print_function
 
-from vivarium.core.tree import Deriver
+from vivarium.core.process import Deriver
 
 
 def divide_condition(compartment):
