+++ vivarium/processes/toxin-antitoxin.py
@@ -4,7 +4,7 @@ import os
 
 import numpy as np
 
-from vivarium.core.tree import Process
+from vivarium.core.process import Process
 from vivarium.core.composition import (
     simulate_process,
     plot_simulation_output,
