+++ vivarium/processes/transcription.py
@@ -10,7 +10,8 @@ import logging as log
 from arrow import StochasticSystem
 
 from vivarium.utils.dict_utils import deep_merge, keys_list
-from vivarium.core.tree import pp, Process
+from vivarium.core.tree import pp
+from vivarium.core.process import Process
 from vivarium.core.composition import process_in_experiment
 from vivarium.states.chromosome import Chromosome, Rnap, Promoter, frequencies, add_merge, toy_chromosome_config
 from vivarium.utils.polymerize import Elongation, build_stoichiometry, template_products
